Move all types to types.ts
